backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.11.2
pydantic-settings==2.8.1
firebase-admin==6.8.0
bcrypt==4.3.0
PyJWT==2.10.1
